split classes
