+++ assets/js/ui/Calculator/CalculatorFirstStep.js
@@ -1,3 +1,6 @@
+import CalculatorSteps from './CalculatorSteps';
+import Equipment from './Equipment';
+
 const renderModel = ($el, carModel) => {
   const blockHTML = `
     <div class="cg-car__name">${carModel.manufacture} ${carModel.name}</div>
@@ -16,71 +19,12 @@ const renderModel = ($el, carModel) => {
   }
 };
 
-class Equipment {
-
-  _node = undefined;
-
-  constructor(id, equipment) {
-    this.id = id;
-    this._equipment = equipment;
-  }
-
-  get name() {
-    return this._equipment.name || '';
-  }
-
-  set name(value) {
-    throw new Error('Read only property "name". Can`t set value:' + value)
-  }
-
-  get range(){
-    return this._equipment.mileageRepeat
-  }
-
-  set range(value) {
-    throw new Error('Read only property "range". Can`t set value:' + value)
-  }
-
-  get works(){
-    return this._equipment.works
-  }
-
-  set works(value) {
-    throw new Error('Read only property "works". Can`t set value:' + value)
-  }
-
-  get node() {
-    if (this._node) {
-      return this._node
-    } else {
-      throw new Error("Node element is undefined")
-    }
-  }
-
-  set node(value) {
-    throw new Error('Read only property "works". Can`t set value:' + value)
-  }
-
-  render(){
-    const wrapper = document.createElement('div');
-    wrapper.innerHTML = `<label class="cg-set__radio">
-                                <input type="radio" name="equipment" value="${this.id}">
-                                <span>${this.name}</span>
-                        </label>`;
-    this._node = wrapper.querySelector('input');
-    this._node.addEventListener('click', () => this.onClick());
-    return wrapper.firstChild
-  }
-
-  onClick() { }
-}
-
-
-class CalculatorFirstStep {
+class CalculatorFirstStep extends CalculatorSteps {
   equipment = undefined;
   range = 0;
 
   constructor(node, equipments) {
+    super(node, equipments);
     this._node = node;
     this.renderEquipments(node, equipments);
   }
@@ -154,17 +98,13 @@ class CalculatorFirstStep {
   showInvalidSelections() {
     const equipmentUnitNode = this._node.querySelector('.costing__col-set').firstElementChild,
           mileageUnitNode = this._node.querySelector('.costing__col-run').firstElementChild;
-    async function highlightNode(node) {
-      node.classList.add('input-error');
-      await setTimeout(() => node.classList.remove('input-error'), 800);
-    }
+
     if (!this.equipment) {
-      highlightNode(equipmentUnitNode).then(() => {});
+      this.highlightNode(equipmentUnitNode).then(() => {});
     }
     if (this.range === 0) {
-      highlightNode(mileageUnitNode).then(() => {});
+      this.highlightNode(mileageUnitNode).then(() => {});
     }
-
   }
 }
