+++ assets/js/ui/Calculator/CalculatorSecondStep.js
@@ -1,152 +1,12 @@
-class CalcItem {
-
-  constructor(id, item) {
-    this.id = id;
-    this._item = item;
-    this._selected = true;
-  }
-
-  get name() {
-    return this._item.name || '';
-  }
-
-  set name(value) {
-    throw new Error('Read only property "name". Can`t set value:' + value)
-  }
-
-  get note() {
-    return this._item.note || '';
-  }
-
-  set note(value) {
-    throw new Error('Read only property "note". Can`t set value:' + value)
-  }
-
-  set totalPrice(value) {
-    throw new Error('Read only property "totalPrice". Can`t set value:' + value)
-  }
-
-  get totalPrice() { }
-
-  get isSelected() { }
-
-  set isSelected(value) {
-    this._selected = value;
-    this._node.checked = value;
-  }
-
-  toString() {
-    const count = +this._item.count,
-          unit = this._item.unit || 'шт',
-          totalPrice = this.totalPrice;
-    return ` ${count > 1 ? count + unit + ' - ' : ''}${totalPrice}`
-  }
-
-  render() {
-    const wrapper = document.createElement('div');
-    wrapper.innerHTML =
-        `<li class="cg-price__item">
-            <div class="cg-price__line">
-                <label class="cg-price__check">
-                  <input type="checkbox" value="${this.id}">
-                  <span>${this.name}</span>
-                </label>
-                <div class="cg-price__cost">
-                    ${this.toString()}
-                    <i class="icon-rub">a</i>
-                </div>
-                <div class="cg-price__info">
-                    <span>${this.note || ''}</span>
-                </div>
-            </div>
-        </li>`;
-    this._node = wrapper.querySelector('input');
-    this._node.addEventListener('click', e => this.isSelected = e.target.checked);
-    return wrapper.firstChild;
-  }
-}
-
-class Work extends CalcItem {
-  _parts = [];
-  constructor(id, item) {
-    super(id, item);
-  }
-
-  get type() {
-    return this._item.type || '';
-  }
-
-  set type(value) {
-    throw new Error('Read only property "type". Can`t set value:' + value)
-  }
-
-  get parts() {
-    return this._parts;
-  }
-
-  set parts(value) {
-    throw new Error('Read only property "parts". Can`t set value:' + value)
-  }
-
-  addPart(part) {
-    this._parts.push(part);
-  }
-
-  set isSelected(status) {
-    super.isSelected = status;
-    this._parts.forEach(part => {
-      part.isSelected = status;
-      part.isDisabled = !status;
-    });
-    this.onChange();
-  }
-
-  get isSelected() {
-    return this._selected;
-  }
-
-  get totalPrice() {
-    return this.isSelected ? +this._item.price : 0;
-  }
-
-
-  onChange() {}
-}
-
-class Part extends CalcItem {
-  constructor(id, item) {
-    super(id, item);
-  }
-
-  get totalPrice() {
-    return this.isSelected ? +this._item.count * +this._item.price : 0;
-  }
-
-  get isSelected() {
-    return this._selected;
-  }
-
-  set isSelected(value) {
-    super.isSelected = value;
-    this.onChange();
-  }
-
-  get isDisabled() {}
-
-  set isDisabled(value) {
-    if (value) this.isSelected = false;
-    this._node.disabled = value;
-  }
-
-  onChange() {}
-
-}
+import Work from './Work';
+import Part from './Part';
 
 class CalculatorSecondStep {
   _works = [];
   range = undefined;
 
   constructor(node, equipment) {
+    this._node = node;
     this.equipment = equipment;
     this.worksNode = node.querySelector('#costing-step_02_works');
     this.recommendationsNode = node.querySelector('#costing-step_02_recommendations');
