+++ assets/js/ui/Calculator/CalculatorThirdStep.js
@@ -1,0 +1,9 @@
+import CalculatorSteps from './CalculatorSteps';
+
+class calculatorThirdStep extends CalculatorSteps {
+  constructor(node) {
+    super();
+  }
+}
+
+export default calculatorThirdStep;
